修改同步消息错误默认行为 (#190)
