+++ items.go
@@ -2,7 +2,6 @@ package openwechat
 
 import (
 	"errors"
-	"strconv"
 )
 
 /*
@@ -89,30 +88,6 @@ func NewUserDetailItemList(members Members) UserDetailItemList {
 		list[index] = item
 	}
 	return list
-}
-
-type SyncCheckResponse struct {
-	RetCode  string
-	Selector string
-}
-
-func (s SyncCheckResponse) Success() bool {
-	return s.RetCode == "0"
-}
-
-func (s SyncCheckResponse) NorMal() bool {
-	return s.Success() && s.Selector == "0"
-}
-
-func (s SyncCheckResponse) Err() error {
-	if s.Success() {
-		return nil
-	}
-	i, err := strconv.ParseInt(s.RetCode, 16, 10)
-	if err != nil {
-		return errors.New("sync check unknown error")
-	}
-	return errors.New(Ret(i).String())
 }
 
 type WebWxSyncResponse struct {
